feat(ui): Add local authentication route
