+++ ui/src/main.tsx
@@ -30,6 +30,7 @@ import WelcomeLocalPasswordRoute from "./routes/welcome-local.password";
 import { CLOUD_API } from "./ui.config";
 import OtherSessionRoute from "./routes/devices.$id.other-session";
 import UpdateRoute from "./routes/devices.$id.update";
+import LocalAuthRoute from "./routes/devices.$id.local-auth";
 
 export const isOnDevice = import.meta.env.MODE === "device";
 export const isInCloud = !isOnDevice;
@@ -86,6 +87,10 @@ if (isOnDevice) {
           path: "update",
           element: <UpdateRoute />,
         },
+        {
+          path: "local-auth",
+          element: <LocalAuthRoute />,
+        },
       ],
     },
 
@@ -137,6 +142,10 @@ if (isOnDevice) {
                 {
                   path: "update",
                   element: <UpdateRoute />,
+                },
+                {
+                  path: "local-auth",
+                  element: <LocalAuthRoute />,
                 },
               ],
             },
